requirements.txt
Flask==2.3.2
python-telegram-bot==20.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0